requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
streamlit==1.39.0
httpx==0.27.2
pandas==2.2.3
numpy==2.1.1
pytest==8.3.3
python-dotenv==1.0.1
